requirements.txt
APScheduler==3.6.3
asgiref==3.5.0
backports.entry-points-selectable==1.1.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.2
dj-rest-auth==2.2.3
Django==4.0.4
django-allauth==0.49.0
django-ckeditor==6.4.0
django-environ==0.8.1
django-filter==21.1
django-js-asset==1.2.2
django-modeltranslation==0.17.5
django-rest-knox==4.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
dynaconf==3.1.8
environ==1.0
filelock==3.0.12
gunicorn==20.1.0
hvac==0.11.2
idna==3.3
knox==0.1.14
loguru==0.6.0
oauthlib==3.2.0
Pillow==9.0.1
platformdirs==2.3.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==1.7.1
pyTelegramBotAPI==3.7.9
python-telegram-bot==13.10
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
tornado==6.1
twilio==6.63.2
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
validators==0.18.2
virtualenv==20.7.2
vue==0.0.1
win32-setctime==1.1.0
